internal: verify ds pods are updated
If a ds is updated and the current method is used to wait for the ds to
be ready, the check will pass but in reality the ds pods will still need
to be updated. this is because the check doesn't ensure the pods are
updated.

Ensure that by adding the condition that the number of up to date pods under the
daemoonset equals to the expected and ready number pods.

Signed-off-by: shajmakh <shajmakh@redhat.com>