e2e: decrease mcp wait interval for Updating condition
For some MCP updates the switch from Updated->Updating->Updated in several
seconds. The default interval for the wait loop of the mcp updates is 30
seconds, which can be problamatic and falky in cases like these.

To avoid this flakiness, decrease the interval time to 2 seconds when
the anticipated new condition is Updating to allow to check the
condition more often.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>