e2e: simplify reboot test 47674
The test goal is to check if updating configurable values under the
numaresourcesoperator CR is done successfully with the anticipated
affects.

The test is flaky in the ci because of mainly 2 issues:
1. waiting improperly for the mcp updates (see commit )
2. not handling the mcp errors on time but instead continuing with the
   test steps

In this commit, we reorder the steps and simplify the test while
retaining the goal of it.

The initial scenario was like this:
1. pick target node
        1.1. label it with new role "mcp-test"
        1.2. remove the worker role
2. create new mcp targeting both roles: mcp-test & worker
        2.1. wait on the mcp mcp-test to get updated
        2.2. wait for mcp worker to start updating and finally reach updated
3. update mcp label on the NROP CR to point to mcp-test
        3.1. verify anticipated results of the CR updates
Now the recovery:
4. undo the NROP CR updates
        4.1. wait for worker mcp to start updating and reach updated
5. delete the mcp mcp-test
6. restore the labels on the target node as follows:
        6.1. remove mcp-test label
        6.2. label it back woth worker

In the scenario above, one can see that already the first step is risky
as it is leaving the target node with unknown mapping for mcp because
mcp `mcp-test` is not yet created.
The second inaccurate thing is that mcp `mcp-test` is targetting two roles:
worker & mcp-test while the intention was to have the target node with 1 label and role only. The difference this does is that if an mcp is
targetting 2 roles the RTE pods will be created on all nodes having
one of these roles, so RTE will still be created on the non-target
nodes, which means mcp worker will not be updated because NROP settings
will still target nodes with worker role, hence step 4 will fail.
The failure however will not stop the test from running, thus it will
continue and do the verifications, and the verification will pass because the test waited long enough on step 4 to update the RTEs on the nodes.
When the test recovers, the recovery is also done incorrectly.
Starting the defers in LIFO order, starting from step 5, it deletes mcp
which holds one machine that has only one role, so after deletion that
one machine will be looking for a machine config but it won't find it
because it was deleted. The other wrong step is that again leaving the
node with no labels (unlabel then label) while it should be done (label
then unlabel). This leaves the environment messed up for the next tests
as mcp is stuck updating.

To fix this we simplify the test (wait for less updates) to make it more stable by following the
below steps:
1. create mcp `mcp-test` targetting one role `mcp-test`
2. pick target node and label it with mcp-test role (keep the old roles)
        2.1. this will trigger mcp update on mcp-test, and it will be
very quick so we allow to check the condition every 2 seconds (see
previous commit)
3. update NROP CR to use mcp-test. this will trigger an update on worker
   mcp because NROP settings will be removed from the non-target nodes.
Nodes associated with mcp-test will not observe an update because NROP is
already configured on them and the only anticipated update would be on
the RTE ds -> restarted with new node-selector
        3.1 wait for mcp update on worker mcp
4. verify nrop cr updates results
Recovery:
5. undo NROP updates
        5.1. wait for mcp update on worker mcp only. remember target
node has worker role.
6. remove mcp-test label off the target node
        6.1 wait for mcp update on worker mcp. this should be quick.
7. delete the mcp-test mcp. no waiting is needed for mcp update.

Signed-off-by: Shereen Haj <shajmakh@redhat.com>