api: types: introduce NodeGroupStatus
Similar there is NodeGroup to set the node group configuration under the
resource's spec, we want to track each node group's status.

So far tracking the status was dependant on having lists of all affected
MachineConfigPools associated with their NodeGroupConfig; list of all
the daemonsets owned by the resources and general operator conditions
and related objects.

As a preparation step to enabling NodeSelector under the node group
(replacing MCP on HCP) we want a way to track the statuses of the node
groups and their configs. Because NodeGroupConfigs in the operator
status were reflected only as part of v1.MachineConfigPool this will not
support our needs. So we want to have a structure that would match for
any selector and a better way to follow on the node-group status,
hence introducing NodeGroupStatus a type that will point to the status
of each NodeGroup in the operator spec. This type includes:
* 1 MCP Name
* 1 Daemonset
* The configuration setting of the RTE on this group

Signed-off-by: Shereen Haj <shajmakh@redhat.com>