api: introduce NodeGroupStatus
So far, tracking the node groups' statuses has been done via the collective operator status, which contains a list of all affected MCPs and their matching RTE daemonsets list.

We keep populating the statuses in these fields for API backward
compatibility and additionally we start reflecting the status per
node group. The relation between the current node group mcp and daemonsets and the new representation is 1:1, and there is no change in the functionality, we are merely providing a new way to gather node group updates under a single entity.

However, this is a required preamble for supporting NodeSelector under NodeGroup, and NodeGroupStatus will be the only place to record and track node group state in HCP (hypershift).

Signed-off-by: Shereen Haj <shajmakh@redhat.com>